requirements.txt
requests
beautifulsoup4
python-dotenv
requests
configparser
pydash
colorlog
prettytable
pypac
tqdm
lxml
